# LoomConnection: speedup scan when using default ordering