Use a slice instead of array of indices to specify default ordering. This allows numpy to return a view of the original array, otherwise it returns a copy.

gh-89

Same trick could be applied to selection, but would require more changes to the code and does not provide much speedup when axis=0.

Attachment 1: Speed comparison of numpy indexing
```
In [1]: a = np.random.random((10000, 10000))

In [2]: %timeit b = a[:, :]                                                                                                                                                 
266 ns ± 7.33 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [3]: %timeit b = a[:, np.fromiter(range(a.shape[1]), dtype=np.int)]                                                                                                      
1.3 s ± 13.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [4]: %timeit b = a[np.fromiter(range(a.shape[1]), dtype=np.int), :]                                                                                                      
476 ms ± 6.31 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)                                                                                                                
```