@JobLeonard 2019-01-24
> This allows numpy to return a view of the original array, otherwise it returns a copy.

I haven't worked with this code in ages, but I can imagine there being code out there that assumes all returned arrays are copies, and this change potentially causing bugs due to mutating values that shouldn't have been mutated. I'm sure @slinnarsson has a better idea of whether this is a problem or not though.