@JobLeonard 2019-01-24
> If I understand correctly, the first line will create a new array, so the second not-copying it should not be a problem.

Ah, so it's removing a _superfluous_ copy. Very nice!